Move start() and stop() as interface methods for CodeExecutor
