+++ python/packages/autogen-core/src/autogen_core/code_executor/_base.py
@@ -5,7 +5,9 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
-from typing import List
+from types import TracebackType
+from typing import List, Optional, Type
+from typing_extensions import Self
 
 from pydantic import BaseModel
 
@@ -30,7 +32,21 @@ class CodeResult:
 
 
 class CodeExecutor(ABC, ComponentBase[BaseModel]):
-    """Executes code blocks and returns the result."""
+    """Executes code blocks and returns the result.
+
+    This is an abstract base class for code executors. It defines the interface
+    for executing code blocks and returning the result. A concrete implementation
+    of this class should be provided to execute code blocks in a specific
+    environment. For example, :class:`~autogen_ext.code_executors.docker.DockerCommandLineCodeExecutor` executes
+    code blocks in a command line environment in a Docker container.
+
+    It is recommended for subclass to be used as a context manager to ensure
+    that resources are cleaned up properly. To do this, implement the
+    :meth:`~autogen_core.code_executor.CodeExecutor.start` and
+    :meth:`~autogen_core.code_executor.CodeExecutor.stop` methods 
+    that will be called when entering and exiting the context manager.
+
+    """
 
     component_type = "code_executor"
 
@@ -54,6 +70,16 @@ class CodeExecutor(ABC, ComponentBase[BaseModel]):
             asyncio.CancelledError: CancellationToken evoked during execution
         """
         ...
+    
+    @abstractmethod
+    async def start(self) -> None:
+        """Start the code executor."""
+        ...
+    
+    @abstractmethod
+    async def stop(self) -> None:
+        """Stop the code executor and release any resources."""
+        ...
 
     @abstractmethod
     async def restart(self) -> None:
@@ -64,3 +90,13 @@ class CodeExecutor(ABC, ComponentBase[BaseModel]):
         This method is called when the agent is reset.
         """
         ...
+
+    async def __aenter__(self) -> Self:
+        await self.start()
+        return self
+
+    async def __aexit__(
+        self, exc_type: Optional[Type[BaseException]], exc_val: Optional[BaseException], exc_tb: Optional[TracebackType]
+    ) -> Optional[bool]:
+        await self.stop()
+        return None
